Refactor email confirmation process for improved modularity and clarity
- Moved email confirmation logic from the main component to a new EmailConfirmationClient component for better separation of concerns.
- Simplified the EmailConfirmation component by utilizing Suspense for loading states.
- Enhanced user feedback during the email verification process with clearer loading and error messages.
- Updated the Supabase client version and added necessary temporary files for project configuration.
