+++ supabase/.temp/cli-latest
@@ -1,1 +1,1 @@
-v2.19.6
+v2.19.7
