+++ supabase/.temp/gotrue-version
@@ -1,0 +1,1 @@
+v2.170.0
