+++ supabase/.temp/pooler-url
@@ -1,0 +1,1 @@
+postgresql://postgres.ijtwvrzkbnfepbfyfrvc:[YOUR-PASSWORD]@aws-0-ap-south-1.pooler.supabase.com:6543/postgres
