+++ supabase/.temp/postgres-version
@@ -1,0 +1,1 @@
+15.8.1.044
